Add new install path to PKG_CONFIG_PATH
cmake-modules-webos installs .pc files to /usr/local/webos/usr/share/pkgconfig
in a standalone build. As a result, a standalone build of LS2 is unable
to locate PmLogLib. This commit adds that path to the PKG_CONFIG_PATH
environment variable. It is a temporary workaround until LS2 is converted
to use cmake-modules-webos.

Open-webOS-DCO-1.0-signed-off-by: Keith Derrick <keith.derrick@palm.com>

Change-Id: Icb09a53290166ae4739ecefc0b3045f1c4629dc9