Deploy luna-send-pub with no access to the private hub
:Release Notes:
Build luna-send-pub on the source code of luna-send.
Let luna-send-pub be world-executable.
Restrict access of luna-send-pub to the public bus only.

:Detailed Notes:
Luna-send must be accessible by third-party developers, but not to
expose WebOS internal (possibly sensitive) data. For that, luna-send
has restricted permissions 0700, and additionally luna-send-pub is
deployed with permissions 0755, but no access to the private hub.

:Testing Performed:
Tested in the emulator and on a device.

:QA Notes:

:Issues Addressed:
[GF-13491] Deploy luna-send-pub with restricted privileges

Open-webOS-DCO-1.0-Signed-off-by: Anatolii Sakhnik <anatolii.sakhnik@lge.com>

Change-Id: Ie1b558bc613a5d4d99dbb2b182f3afb24ca715bd
Reviewed-on: https://g2g.palm.com/3205
Reviewed-by: Build Verification
Reviewed-by: Anatolii Sakhnik <anatolii.sakhnik@lge.com>
Tested-by: Anatolii Sakhnik <anatolii.sakhnik@lge.com>
Reviewed-by: Alekseyev Oleksandr <alekseyev.oleksandr@lge.com>
Tested-by: Alekseyev Oleksandr <alekseyev.oleksandr@lge.com>