ls-hubd_watchdog to reboot system on ls-hub crash
:Release Notes:
If ls-hubd_private or ls-hubd_public crash, the only known remedy is to
reboot. This watchdog triggers on stopping ls-hubd_* and if the stopping
is not a desired event, it will initiate a reboot.

:Detailed Notes:
If ls-hubd_* are stopped by a user or as part of a normal shutdown, the
env var RESULT presented by Upstart will be "ok", or if there is a
normal reboot or shutdown occurring (triggered by RUNLEVEL 0,1,6), then
a watchdog reboot is not needed.  If it is needed, watchdog will add
LD_HUBD_CRASH=<name of the stopping job> ("ls-hubd_public" or
"ls-hubd_private") to the global Upstart env so that reboot knows how to
handle the process.

:Testing Performed:
Many reboots were initated via the reboot and shutdown commands and also
by killing -ABRT ls-hubd_*
Also ran a miniBAT for good measure.

:QA Notes:

:Issues Addressed:
[GF-45875] ls-hubd_p*_watchdog not working

Open-webOS-DCO-1.0-Signed-off-by: Ed Chejlava <ed.chejlava@lge.com>
Change-Id: I8260e3462354ccbc5a7bcbce6ded1ee85635e2dc
Reviewed-on: https://g2g.palm.com/3978
Reviewed-by: DCO Verification
Reviewed-by: Keith Derrick <keith.derrick@lge.com>
Tested-by: Keith Derrick <keith.derrick@lge.com>