Fix watchdog, add oom exclusion to ls-hubd_p*.conf
:Release Notes:
Current installed ls-hubd_watchdog doesn't work - it has a spurious
paren.  Add "oom score never" to both ls-hubd jobs as we do not
want them to ever be killed to free up memory.

:Detailed Notes:
ls-hubd watchdog has not worked since webOS 3.0.5.  It was missing bits
and pieces, and until this fix had at least one syntax error in it.
Also, we've seen crashes that seem to be occurring due to init killing
ls-hubd* due to an OOM condition.  The "oom score" is fairly new in
Upstart so I'll try setting it to "oom score never" to see if that
prevents the crashes.

:Testing Performed:
kill (*) ls-hubd_* and verify that the reboot happens.
miniBAT results attached to CCC.
(*): Some signals - HUP, QUIT, INT - end the process the "normal"
way and won't work to test the watchdog.  I found that TRAP, ABRT,
KILL, SEGV all worked to simulate a crash.

:QA Notes:

:Issues Addressed:
[GF-52636] ls2 watchdog script incorrectly installed
[GF-52116] Loading Settings panel during youtube play in browser
           crashes system
[GF-52624] CCC: luna-service2=166

Open-webOS-DCO-1.0-Signed-off-by: Ed Chejlava <ed.chejlava@lge.com>

Change-Id: I357e75a1cdb6aeb5a463496f47013e5e94eae3c1
Reviewed-on: https://g2g.palm.com/4037
Reviewed-by: DCO Verification
Reviewed-by: Ed Chejlava <ed.chejlava@lge.com>
Tested-by: Ed Chejlava <ed.chejlava@lge.com>