Reset /proc/PID/oom_* for ls-hubd spawned procs
:Release Notes:
A recent change to ls-hubd's Upstart jobs files adding 'oom score never'
has the unintended side-effect of giving that same protection from the
oom killer to all spawned jobs (i.e. all dynamic services).  This fix
handles that by sending "0" to oom_adj for each process that gets
spawned.

:Detailed Notes:
While 'normal' processes return "0" when you 'cat /proc/PID/oom_adj',
oom_score or oom_score_adj, setting 'oom score never' in the Upstart
job results in values of -17, 0 and -1000 for the process started
by that job: in this case ls-hubd (private and public).  This is what
we wanted when we added that line, as if the oom killer kills ls-hubd,
it is effectively killing webOS.
The side effect that all dynamic processes (which are spawned by
ls-hubd) will inherit those settings was not expected.  It is "bad"
for those processes to be immune to the oom killer as they are not
critical to the device's operation.
This fix simply gets the PID from a successfully spawned process and
in sequence opens, write "0" to and closes /proc/PID/oom_adj which
will; cause the kernel to resetting all oom* to 'normal' values.

:Testing Performed:
miniBAT was run on verification build svl.multilayergoldfinger-176,
and manual testing has confirmed that the three /proc/PID/oom_* all
return "0" for procs started by ls-hubd private (and that ls-hubd
still has the previous -17, 0, -1000 values.
miniBAT results attached to CCC.

:QA Notes:

:Issues Addressed:
[GF-54748] oom_score_adj of dynamic services should not be -1000
[GF-55294] CCC: luna-service2=169

Open-webOS-DCO-1.0-Signed-off-by: Ed Chejlava <ed.chejlava@lge.com>

Change-Id: I01745d0e5a0cb08ae2f4e6f20413b2fa0ad08ed7
Reviewed-on: https://g2g.palm.com/4185
Reviewed-by: DCO Verification
Reviewed-by: Ed Chejlava <ed.chejlava@lge.com>
Tested-by: Ed Chejlava <ed.chejlava@lge.com>