Add -w option to luna-send
:Release Notes:
Add a command line option to luna-send to allow scripts to control the
length of time it waits for a command to complete.

:Detailed Notes:
The delay was hard-coded at 300ms. During shutdown, with multiple
invocations, this can add over 1 second to the total shutdown time.

The -w option allows a different delay value to be specified (in
milliseconds).

:Testing Performed:
Checked to ensure that luna-send still functioned correctly, then used
the time command with various values of -w to ensure that it had the
expected results.

:QA Notes:

:Issues Addressed:
[GF-60639] Speed up shutdown by limiting luna-send's timeout

Open-webOS-DCO-1.0-Signed-off-by: Keith Derrick <keith.derrick@lge.com>

Change-Id: I97d5a88563536a79deabfe16539eb24b139b4cdb
Reviewed-on: https://g2g.palm.com/4678
Reviewed-by: DCO Verification
Reviewed-by: Herb Kuta <herb.kuta@lge.com>
Tested-by: Herb Kuta <herb.kuta@lge.com>
Reviewed-by: Keith Derrick <keith.derrick@lge.com>
Tested-by: Keith Derrick <keith.derrick@lge.com>