Prevent loss of calls
:Release Notes:
Prevent loss of calls
Prevent usage of already released object

:Detailed Notes:
The first issue.
When server is shutting down, it can be a situation when the last message
"shutdown message" is not transmitted completely and a client receives
SIG_PIPE. As a result, already sent messages become lost and the client does
not receive reply.
The second issue. An object of _LSTransportClient was used without increasing
of a reference counter. As a result the object was released but still was
used in another thread.

:Testing Performed:
Unit tests.
Smoke on the target.

:QA Notes:

:Issues Addressed
[GF-61218] LSCallOneReply sometimes doesn't reply

Open-webOS-DCO-1.0-Signed-off-by: Alekseyev Oleksandr<alekseyev.oleksandr@lge.com>

Change-Id: I474b305423524d639134d192f7c55fd8f4b8e7af
Reviewed-on: https://g2g.palm.com/4880
Reviewed-by: DCO Verification
Reviewed-by: Alekseyev Oleksandr <alekseyev.oleksandr@lge.com>
Tested-by: Alekseyev Oleksandr <alekseyev.oleksandr@lge.com>
Reviewed-by: Anatolii Sakhnik <anatolii.sakhnik@lge.com>