Fix crash in LSUnregister() with dynamic peer
:Release Notes:
Fix crash in LSUnregister() when connecting to a dynamic service.

:Detailed Notes:
Enhance hub runner to launch dynamic services.  When LSUnregister() is
called, transport and callmap are deinitialized.  If pending messages
to a dynamic service are still in the queue, failure notification is
sent back to the client. The callmap must outlive the transport for
this reason.

:Testing Performed:
make check && make test

:QA Notes:

:Issues Addressed:
[BHV-3965] luna-send crashes at location /lib/libc.so.6(gsignal+0x3c)

Open-webOS-DCO-1.0-Signed-off-by: Anatolii Sakhnik <anatolii.sakhnik@lge.com>

Change-Id: I460262fa3c383e9755b3e74853eb5de006249387
Reviewed-on: https://g2g.palm.com/5212
Reviewed-by: DCO Verification
Reviewed-by: Anatolii Sakhnik <anatolii.sakhnik@lge.com>
Tested-by: Anatolii Sakhnik <anatolii.sakhnik@lge.com>
Reviewed-by: Nikolay Orliuk <nikolay.orliuk@lge.com>