Use GSList instead of obscured GQueue
:Release Notes:

:Detailed Notes:
List of pattern specs used to be kept in GQueue for inbound and
outbound permissions. We need a more transparent data structure to
compare two lists for equality efficiently.
We're going to keep inbound and outbound permissions in sorted lists.
Then we'd be able to compare them element-wise in a single pass.

:Testing Performed:
Build and automatic tests

:QA Notes:

:Issues Addressed:
[BHV-9332] Multiple ls-hub errors registering media pipeline

Open-webOS-DCO-1.0-Signed-off-by: Anatolii Sakhnik <anatolii.sakhnik@lge.com>

Change-Id: I125298f7bcfefaa662ec4a50b78ab49a3830b28d
Reviewed-on: https://g2g.palm.com/5913
Reviewed-by: DCO Verification
Reviewed-by: Anatolii Sakhnik <anatolii.sakhnik@lge.com>
Tested-by: Anatolii Sakhnik <anatolii.sakhnik@lge.com>
Reviewed-by: Sergiy Kryvonos <sergiy.kryvonos@lge.com>