Wait for populate_volatile before creating sockets
:Release Notes:

Start the private hub only after the populate_volatile task has
completed to ensure /tmp is properly set up.

:Detailed Notes:
During regression testing we found that occasionally LS2 did not
seem to start up properly. On investigation, only the owner had
write access to the private hub socket in /tmp.

The populate_volatiles script converts /tmp to a link to
/var/volatiles/tmp, and this was happening in parallel with ls2
starting up.

This problem only occurred on the very first boot after an erase and
reflash of a device. Adjusting the startup conditions for LS2 appears
to cure the problem.

:Testing Performed:
Ran a test loop which repeatedly erased, reflashed, and booted a
device - checking for a failure.

:QA Notes:

:Issues Addressed:

[CTO-431] LS2 does not come up on very first boot after flash

Open-webOS-DCO-1.0-Signed-off-by: Keith Derrick <keith.derrick@lge.com>

Change-Id: I0a59392204f84032b96388fdcc25615fc0c32147
Reviewed-on: https://g2g.palm.com/6452
Reviewed-by: DCO Verification
Reviewed-by: Ed Chejlava <ed.chejlava@lge.com>
Reviewed-by: Gayathri Srinivasan <gayathri.srinivasan@lge.com>
Reviewed-by: Keith Derrick <keith.derrick@lge.com>
Tested-by: Keith Derrick <keith.derrick@lge.com>